requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.15.1
django-filter==24.2
numpy==2.2.6
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
